.cursor/skills/pm/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-frontmatter>=1.0.0
pyyaml>=6.0
